pipelines/segmentation/legend_and_map/requirements.txt
# requirements file for legend_and_map segmentation pipeline
# (Additional requirements under corresponding 'tasks' sub-dir)
Flask==3.0.0
python-dotenv==1.0.0